refactor: fields cleanup
