+++ packages/dm-core-plugins/src/form/fields/BooleanField/BooleanField.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, screen, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { Form } from '../components/Form'
-import { mockBlueprintGet, wrapper } from '../test-utils'
+import { Form } from '../../components/Form'
+import { mockBlueprintGet, wrapper } from '../../test-utils'
 
 describe('BooleanField', () => {
   afterEach(() => {
