+++ packages/dm-core-plugins/src/form/fields/BooleanField/BooleanField.tsx
@@ -1,8 +1,8 @@
 import { Controller, useFormContext } from 'react-hook-form'
-import { useRegistryContext } from '../context/RegistryContext'
-import { getWidget } from '../context/WidgetContext'
-import { TField } from '../types'
-import { getDisplayLabel } from '../utils/getDisplayLabel'
+import { useRegistryContext } from '../../context/RegistryContext'
+import { getWidget } from '../../context/WidgetContext'
+import { TField } from '../../types'
+import { getDisplayLabel } from '../../utils/getDisplayLabel'
 
 export const BooleanField = (props: TField) => {
   const { namePath, uiAttribute, attribute } = props
