+++ packages/dm-core-plugins/src/form/fields/ObjectField/ObjectField.test.tsx
@@ -6,9 +6,9 @@ import {
   waitFor,
 } from '@testing-library/react'
 import React from 'react'
-import { Form } from '../components/Form'
-import { mockBlueprintGet, wrapper } from '../test-utils'
-import { TFormProps } from '../types'
+import { Form } from '../../components/Form'
+import { mockBlueprintGet, wrapper } from '../../test-utils'
+import { TFormProps } from '../../types'
 
 afterEach(() => cleanup())
