+++ packages/dm-core-plugins/src/form/fields/ObjectField/templates/ObjectStorageUncontainedTemplate.tsx
@@ -5,16 +5,16 @@ import {
 } from '@development-framework/dm-core'
 import { useState } from 'react'
 import { useFormContext } from 'react-hook-form'
-import { OpenObjectButton } from '../components/OpenObjectButton'
-import RemoveObject from '../components/RemoveObjectButton'
-import { useRegistryContext } from '../context/RegistryContext'
-import { TObjectTemplate } from '../types'
-import FormTemplate from './shared/FormTemplate'
+import { OpenObjectButton } from '../../../components/OpenObjectButton'
+import RemoveObject from '../../../components/RemoveObjectButton'
+import { useRegistryContext } from '../../../context/RegistryContext'
+import FormTemplate from '../../../templates/shared/FormTemplate'
 import {
   getCanOpenOrExpand,
   getExpandViewConfig,
   getOpenViewConfig,
-} from './shared/utils'
+} from '../../../templates/shared/utils'
+import { TObjectTemplate } from '../../../types'
 
 export const ObjectStorageUncontainedTemplate = (props: TObjectTemplate) => {
   const { namePath, attribute, uiAttribute } = props
