+++ packages/dm-core-plugins/src/form/fields/StringField/StringField.tsx
@@ -1,8 +1,8 @@
 import { Controller, useFormContext } from 'react-hook-form'
-import { useRegistryContext } from '../context/RegistryContext'
-import { getWidget } from '../context/WidgetContext'
-import { TField } from '../types'
-import { getDisplayLabel } from '../utils/getDisplayLabel'
+import { useRegistryContext } from '../../context/RegistryContext'
+import { getWidget } from '../../context/WidgetContext'
+import { TField } from '../../types'
+import { getDisplayLabel } from '../../utils/getDisplayLabel'
 
 export const StringField = (props: TField) => {
   const { namePath, uiAttribute, attribute } = props
