+++ packages/dm-core-plugins/src/form/fields/index.tsx
@@ -1,15 +1,7 @@
-import ArrayField from './ArrayField'
-import { AttributeField } from './AttributeField'
-import { BooleanField } from './BooleanField'
-import { NumberField } from './NumberField'
-import { ObjectField } from './ObjectField'
-import { StringField } from './StringField'
-
-export default {
-  ArrayField,
-  BooleanField,
-  NumberField,
-  ObjectField,
-  AttributeField,
-  StringField,
-}
+export * from './ArrayField/ArrayField'
+export * from './AttributeField'
+export * from './BinaryField/BinaryField'
+export * from './BooleanField/BooleanField'
+export * from './NumberField/NumberField'
+export * from './ObjectField/ObjectField'
+export * from './StringField/StringField'
